Refactored command functions into separate files. Ran Prettier on many files.
