+++ src/lib/commands.ts
@@ -1,27 +1,44 @@
 import * as vscode from "vscode";
+
 import { getEol } from "./env";
-import {
-  ExtensionContext,
-  QuickPickOptions,
-  QuickPickItem,
-  Selection,
-  TextEditor,
-} from "vscode";
+
+import { ExtensionContext, QuickPickOptions, QuickPickItem } from "vscode";
+
 import { addTable } from "./tables";
+
+import { toggleAdministration } from "./commands/toggle-administration";
+import { toggleAvailability } from "./commands/toggle-availability";
+import { toggleBold } from "./commands/toggle-bold";
+import { toggleBullets } from "./commands/toggle-bullets";
+import { toggleCaution } from "./commands/toggle-caution";
+import { toggleCheckboxes } from "./commands/toggle-checkboxes";
+import { toggleCitations } from "./commands/toggle-citations";
+import { toggleCodeBlock, toggleInlineCode } from "./commands/toggle-code";
+import { toggleDNL } from "./commands/toggle-dnl";
+import { toggleError } from "./commands/toggle-error";
 import {
-  surroundSelection,
-  surroundBlockSelection,
-  isAnythingSelected,
-  isBlockMatch,
-  replaceBlockSelection,
-  prefixLines,
-  getSurroundingWord,
-  replaceSelection,
-  getLineSelection,
-  isMatch,
-  getBlockSelection,
-  reSelect,
-} from "./editorHelpers";
+  toggleTitleH1,
+  toggleTitleH2,
+  toggleTitleH3,
+  toggleTitleH4,
+  toggleTitleH5,
+  toggleTitleH6,
+} from "./commands/toggle-headers";
+import { toggleImage } from "./commands/toggle-image";
+import { toggleImportant } from "./commands/toggle-important";
+import { toggleInfo } from "./commands/toggle-info";
+import { toggleItalic } from "./commands/toggle-italic";
+import { toggleLink } from "./commands/toggle-link";
+import { toggleMoreLikeThis } from "./commands/toggle-morelikethis";
+import { toggleNote } from "./commands/toggle-note";
+import { toggleNumberList } from "./commands/toggle-numberlist";
+import { togglePrerequisites } from "./commands/toggle-prerequisites";
+import { toggleStrikethrough } from "./commands/toggle-strikethrough";
+import { toggleSuccess } from "./commands/toggle-success";
+import { toggleTip } from "./commands/toggle-tip";
+import { toggleUIControl } from "./commands/toggle-uicontrol";
+import { toggleVideo } from "./commands/toggle-video";
+import { toggleWarning } from "./commands/toggle-warning";
 
 interface CommandItem extends QuickPickItem {
   label: string;
@@ -283,7 +300,7 @@ const _commands: Command[] = [
   ),
 ];
 
-let newLine = getEol();
+export const newLine = getEol();
 
 export function register(context: ExtensionContext) {
   _commands.map((cmd) => {
@@ -311,514 +328,5 @@ function showCommandPalette() {
     });
 }
 
-const wordMatch: string = "[A-Za-z\\u00C0-\\u017F]";
-
-interface BoldExpressions {
-  [idx: string]: RegExp;
-}
-
-interface ItalicExpressions {
-  [idx: string]: RegExp;
-}
-
-const toggleBoldExpressions: BoldExpressions = {
-  "**": new RegExp(`\\*{2}${wordMatch}*\\*{2}|${wordMatch}+`),
-  __: new RegExp(`_{2}${wordMatch}*_{2}|${wordMatch}+`),
-};
-
-const toggleItalicExpressions: ItalicExpressions = {
-  "_": new RegExp(`\\_{1}${wordMatch}*\\_{1}|${wordMatch}+`),
-  __: new RegExp(`\\{2}${wordMatch}*\\{2}|${wordMatch}+`),
-};
-
-function toggleBold(): void | Thenable<void> | Thenable<boolean> {
-  const marker: string | undefined = vscode.workspace
-    .getConfiguration("markdownShortcuts.bold")
-    .get("marker");
-  if (!marker) {
-    return;
-  }
-
-  return surroundSelection(marker, marker, toggleBoldExpressions[marker]);
-}
-
-function toggleItalic(): void | Thenable<void> | Thenable<boolean> {
-  const marker: string | undefined = vscode.workspace
-    .getConfiguration("markdownShortcuts.italics")
-    .get("marker");
-  if (!marker) {
-    return;
-  }
-  const pattern: RegExp = new RegExp(`\\${marker}?${wordMatch}*\\${marker}?`);
-
-  return surroundSelection(marker, marker, toggleItalicExpressions[marker]);
-}
-
-const toggleStrikethroughPattern: RegExp = new RegExp(
-  "~{2}" + wordMatch + "*~{2}|" + wordMatch + "+"
-);
-function toggleStrikethrough() {
-  return surroundSelection("~~", "~~", toggleStrikethroughPattern);
-}
-
-const startingBlock: string = "```" + newLine;
-const endingBlock: string = newLine + "```";
-const codeBlockWordPattern: RegExp = new RegExp(
-  `${startingBlock}.+${endingBlock}|.+`,
-  "gm"
-);
-function toggleCodeBlock() {
-  return surroundBlockSelection(
-    startingBlock,
-    endingBlock,
-    codeBlockWordPattern
-  );
-}
-
-const toggleInlineCodePattern = new RegExp(
-  "`" + wordMatch + "*`|" + wordMatch + "+"
-);
-function toggleInlineCode() {
-  return surroundSelection("`", "`", toggleInlineCodePattern);
-}
-
-const headerWordPattern = /#{1,6} .+|.+/;
-function toggleTitleH1() {
-  return surroundSelection("# ", "", headerWordPattern);
-}
-
-function toggleTitleH2() {
-  return surroundSelection("## ", "", headerWordPattern);
-}
-
-function toggleTitleH3() {
-  return surroundSelection("### ", "", headerWordPattern);
-}
-
-function toggleTitleH4() {
-  return surroundSelection("#### ", "", headerWordPattern);
-}
-
-function toggleTitleH5() {
-  return surroundSelection("##### ", "", headerWordPattern);
-}
-
-function toggleTitleH6() {
-  return surroundSelection("###### ", "", headerWordPattern);
-}
-
-const addBullets: RegExp = /^(\s*)(.+)$/gm;
-function toggleBullets() {
-  const marker = vscode.workspace
-    .getConfiguration("markdownShortcuts.bullets")
-    .get("marker");
-
-  if (!isAnythingSelected()) {
-    return surroundSelection(
-      marker + " ",
-      "",
-      new RegExp("\\" + marker + " .+|.+")
-    );
-  }
-
-  const hasBullets = new RegExp("^(\\s*)\\" + marker + " (.*)$", "gm");
-
-  if (isBlockMatch(hasBullets)) {
-    return replaceBlockSelection((text) => text.replace(hasBullets, "$1$2"));
-  } else {
-    return replaceBlockSelection((text) =>
-      text.replace(addBullets, "$1" + marker + " $2")
-    );
-  }
-}
-
-interface LineNums {
-  [index: string]: number;
-}
-
-const hasNumbers: RegExp = /^(\s*)[0-9]\.+ (.*)$/gm;
-const addNumbers: RegExp = /^(\n?)(\s*)(.+)$/gm;
-function toggleNumberList() {
-  if (!isAnythingSelected()) {
-    return surroundSelection("1. ", "");
-  }
-
-  if (isBlockMatch(hasNumbers)) {
-    return replaceBlockSelection((text) => text.replace(addNumbers, "$1$2"));
-  } else {
-    const lineNums: LineNums = {};
-    return replaceBlockSelection((text) =>
-      text.replace(addNumbers, (match, newline, whitespace, line) => {
-        if (!lineNums[whitespace]) {
-          lineNums[whitespace] = 1;
-        }
-        return newline + whitespace + lineNums[whitespace]++ + ". " + line;
-      })
-    );
-  }
-}
-
-const hasCheckboxes: RegExp = /^(\s*)- \[[ x]{1}\] (.*)$/gim;
-const addCheckboxes: RegExp = /^(\s*)(.+)$/gm;
-function toggleCheckboxes() {
-  if (!isAnythingSelected()) {
-    return surroundSelection("- [ ] ", "", /- \[[ x]{1}\] .+|.+/gi);
-  }
-
-  if (isBlockMatch(hasCheckboxes)) {
-    return replaceBlockSelection((text) => text.replace(hasCheckboxes, "$1$2"));
-  } else {
-    return replaceBlockSelection((text) =>
-      text.replace(addCheckboxes, "$1- [ ] $2")
-    );
-  }
-}
-
-const hasCitations = /^(\s*)> (.*)$/gim;
-const addCitations = /^(\s*)(.*)$/gm;
-
-function toggleCitations() {
-  if (!isAnythingSelected()) {
-    return surroundSelection("> ", "", /> .+|.+/gi);
-  }
-
-  if (isBlockMatch(hasCitations)) {
-    return replaceBlockSelection((text) => text.replace(hasCitations, "$1$2"));
-  } else {
-    return prefixLines("> ");
-  }
-}
-
-interface LinkUrl {
-  text: string;
-  url: string | undefined;
-}
-
-function getLinkText(selection: Selection): Thenable<string | undefined> {
-  if (selection.isEmpty) {
-    return vscode.window.showInputBox({
-      prompt: "Image alt text",
-    });
-  }
-
-  return Promise.resolve("");
-}
-
-function getLinkUrl(
-  linkText: string | undefined
-): LinkUrl | Thenable<LinkUrl> | void {
-  if (linkText === null || linkText === undefined) {
-    return;
-  }
-
-  return vscode.window
-    .showInputBox({
-      prompt: "Image URL",
-    })
-    .then((url) => {
-      return { text: linkText, url: url };
-    });
-}
-
-function addTags(options: LinkUrl | void): void | Thenable<void> {
-  if (!options || !options.url) {
-    return;
-  }
-  surroundSelection("![" + options.text, "](" + options.url + ")");
-}
-
-const markdownLinkRegex: RegExp = /^\[.+\]\(.+\)$/;
-const urlRegex: RegExp = /^(http[s]?:\/\/.+|<http[s]?:\/\/.+>)$/;
-const markdownLinkWordPattern: RegExp = new RegExp(
-  "[.+](.+)|" + wordMatch + "+"
-);
-function toggleLink():
-  | Thenable<LinkUrl>
-  | void
-  | Thenable<void>
-  | Thenable<boolean> {
-  const editor: TextEditor | undefined = vscode.window.activeTextEditor;
-  if (!editor) {
-    return;
-  }
-  let selection: Selection = editor.selection;
-
-  if (!isAnythingSelected()) {
-    const withSurroundingWord = getSurroundingWord(
-      editor,
-      selection,
-      markdownLinkWordPattern
-    );
-
-    if (withSurroundingWord) {
-      selection = editor.selection = withSurroundingWord;
-    }
-  }
-
-  if (isAnythingSelected()) {
-    if (isMatch(markdownLinkRegex)) {
-      //Selection is a MD link, replace it with the link text
-      return replaceSelection((text) => {
-        const mdLink: RegExpMatchArray | null = text.match(/\[(.+)\]/);
-        return mdLink ? mdLink[1] : text;
-      });
-    }
-
-    if (isMatch(urlRegex)) {
-      //Selection is a URL, surround it with angle brackets
-      return surroundSelection("<", ">");
-    }
-  }
-
-  return getLinkText(selection).then(getLinkUrl).then(addTags);
-}
-
-const markdownImageRegex: RegExp = /!\[.*\]\((.+)\)/;
-function toggleImage() {
-  const editor: TextEditor | undefined = vscode.window.activeTextEditor;
-  if (!editor) {
-    return;
-  }
-  let selection: Selection = editor.selection;
-
-  if (isAnythingSelected()) {
-    if (isMatch(markdownImageRegex)) {
-      //Selection is a MD link, replace it with the link text
-      return replaceSelection((text) => {
-        const textMatch: RegExpMatchArray | null = text.match(
-          markdownImageRegex
-        );
-        return textMatch ? textMatch[1] : "";
-      });
-    }
-
-    if (isMatch(urlRegex)) {
-      return vscode.window
-        .showInputBox({
-          prompt: "Image alt text",
-        })
-        .then((text) => {
-          if (text === null) {
-            return;
-          }
-          replaceSelection((url) => "![" + text + "](" + url + ")");
-        });
-    }
-  }
-  return getLinkText(selection).then(getLinkUrl).then(addTags);
-}
-
-const startingNote: string = `>[!NOTE]${newLine}>${newLine}>`;
-const endingNote: string = newLine;
-const noteBlockWordPattern: RegExp = new RegExp(
-  startingNote + ".+" + endingNote + "|.+",
-  "gm"
-);
-function toggleNote() {
-  return surroundBlockSelection(startingNote, endingNote, noteBlockWordPattern);
-}
-
-const startingTip: string = ">[!TIP]" + newLine + ">" + newLine + ">";
-const endingTip: string = newLine;
-const tipBlockWordPattern: RegExp = new RegExp(
-  startingTip + ".+" + endingTip + "|.+",
-  "gm"
-);
-function toggleTip() {
-  return surroundBlockSelection(startingTip, endingTip, tipBlockWordPattern);
-}
-
-const startingCaution: string = ">[!CAUTION]" + newLine + ">" + newLine + ">";
-const endingCaution: string = newLine;
-const cautionBlockWordPattern: RegExp = new RegExp(
-  startingCaution + ".+" + endingCaution + "|.+",
-  "gm"
-);
-function toggleCaution() {
-  return surroundBlockSelection(
-    startingCaution,
-    endingCaution,
-    cautionBlockWordPattern
-  );
-}
-
-const startingWarning: string = ">[!WARNING]" + newLine + ">" + newLine + ">";
-const endingWarning = newLine;
-const warningBlockWordPattern = new RegExp(
-  startingWarning + ".+" + endingWarning + "|.+",
-  "gm"
-);
-function toggleWarning() {
-  return surroundBlockSelection(
-    startingWarning,
-    endingWarning,
-    warningBlockWordPattern
-  );
-}
-
-const startingImportant = ">[!IMPORTANT]" + newLine + ">" + newLine + ">";
-const endingImportant = newLine;
-const importantBlockWordPattern = new RegExp(
-  startingImportant + ".+" + endingImportant + "|.+",
-  "gm"
-);
-function toggleImportant() {
-  return surroundBlockSelection(
-    startingImportant,
-    endingImportant,
-    importantBlockWordPattern
-  );
-}
-
-const startingAvailability: string = `>[!AVAILABILITY]${newLine}>${newLine}>`;
-const endingAvailability: string = newLine;
-const availabilityBlockWordPattern: RegExp = new RegExp(
-  startingAvailability + ".+" + endingAvailability + "|.+",
-  "gm"
-);
-function toggleAvailability() {
-  return surroundBlockSelection(startingAvailability, endingAvailability, availabilityBlockWordPattern);
-}
-
-const startingAdministration: string = `>[!ADMINISTRATION]${newLine}>${newLine}>`;
-const endingAdministration: string = newLine;
-const administrationBlockWordPattern: RegExp = new RegExp(
-  startingAdministration + ".+" + endingAdministration + "|.+",
-  "gm"
-);
-function toggleAdministration() {
-  return surroundBlockSelection(startingAdministration, endingAdministration, administrationBlockWordPattern);
-}
-
-const startingPrerequisites: string = `>[!PREREQUISITES]${newLine}>${newLine}>`;
-const endingPrerequisites: string = newLine;
-const prerequisitesBlockWordPattern: RegExp = new RegExp(
-  startingPrerequisites + ".+" + endingPrerequisites + "|.+",
-  "gm"
-);
-function togglePrerequisites() {
-  return surroundBlockSelection(startingPrerequisites, endingPrerequisites, prerequisitesBlockWordPattern);
-}
-
-const startingError: string = `>[!ERROR]${newLine}>${newLine}>`;
-const endingError: string = newLine;
-const errorBlockWordPattern: RegExp = new RegExp(
-  startingError + ".+" + endingError + "|.+",
-  "gm"
-);
-function toggleError() {
-  return surroundBlockSelection(startingError, endingError, errorBlockWordPattern);
-}
-
-const startingInfo: string = `>[!INFO]${newLine}>${newLine}>`;
-const endingInfo: string = newLine;
-const infoBlockWordPattern: RegExp = new RegExp(
-  startingInfo + ".+" + endingInfo + "|.+",
-  "gm"
-);
-function toggleInfo() {
-  return surroundBlockSelection(startingInfo, endingInfo, infoBlockWordPattern);
-}
-
-const startingSuccess: string = `>[!SUCCESS]${newLine}>${newLine}>`;
-const endingSuccess: string = newLine;
-const successBlockWordPattern: RegExp = new RegExp(
-  startingSuccess + ".+" + endingSuccess + "|.+",
-  "gm"
-);
-function toggleSuccess() {
-  return surroundBlockSelection(startingSuccess, endingSuccess, successBlockWordPattern);
-}
-
-const startingMoreLikeThis =
-  ">[!MORELIKETHIS]" + newLine + '>' + newLine;
-const endingMoreLikeThis = newLine;
-const moreLikeThisBlockWordPattern = new RegExp(
-  startingMoreLikeThis + ".+" + endingMoreLikeThis + "|.+",
-  "gm"
-);
-function toggleMoreLikeThis(): Thenable<boolean | void> {
-  return surroundBlockSelection(
-    startingMoreLikeThis,
-    endingMoreLikeThis,
-    moreLikeThisBlockWordPattern
-  );
-}
-
-function addTagsToVideo(url: string): Thenable<boolean> {
-  return surroundSelection(`>[!VIDEO](${url})`, '');
-}
-
-function getLinkUrlToVideo(
-): Thenable<string> {
-  return vscode.window
-    .showInputBox({
-      prompt: "Video URL",
-    })
-    .then((url) => {
-      return url || '';
-    });
-}
-
-const markdownVideoRegex: RegExp = /^>\[\!VIDEO\]\(.+\).*/;
-const videoUrlRegex: RegExp = /https?:\/\/(www\.)?[-a-zA-Z0-9@:%._\+~#=]{1,256}\.[a-zA-Z0-9()]{1,6}\b([-a-zA-Z0-9()@:%_\+.~#?&//=]*)/;
-
-function toggleVideo(): Thenable<boolean> {
-  const editor: TextEditor | undefined = vscode.window.activeTextEditor;
-  if (!editor) {
-    return Promise.reject('No text editor available');
-  }
-
-  // VIDEO tags work on the current line, so ignore the selection and select the whole line.
-  let selection: Selection = editor.selection = getLineSelection() || editor.selection;
-
-  // If anything is selected, look for an existing VIDEO tag.
-  if (isAnythingSelected()) {
-    if (isMatch(markdownVideoRegex)) {
-      //Selection is a MD link, replace it with the link text
-      return replaceSelection((text) => {
-        const videoUrl: RegExpMatchArray | null = text.match(/\((.+)\)/); // Match everything in parentheses
-        return videoUrl ? videoUrl[1] : text;
-      });
-    }
-
-    if (isMatch(videoUrlRegex)) {
-      return replaceSelection((text) => {
-        const videoUrl: RegExpMatchArray | null = text.match(videoUrlRegex);
-        if (!(videoUrl && videoUrl.input)) {
-          return text; // Should never happen because of the isMatch condition
-        } else {
-          if (videoUrl && videoUrl[0].length < videoUrl.input.length && videoUrl.index) {
-            return `>[!VIDEO](${videoUrl[0]}) ${videoUrl.input}`;
-          } else {
-            return `>[!VIDEO](${videoUrl[0]})`;
-          }
-        }
-      });
-    }
-  }
-
-  const linkToVideo = getLinkUrlToVideo();
-
-  return linkToVideo.then(
-    (linkObj) => {
-      if (linkObj) {
-        return addTagsToVideo(linkObj);
-      } else { return Promise.reject('No URL provided.'); };
-    });
-
-}
-
-const toggleDNLPattern: RegExp = new RegExp(
-  "[!DNL" + wordMatch + "]" + wordMatch + "+"
-);
-function toggleDNL() {
-  return surroundSelection("[!DNL ", "]", toggleDNLPattern);
-}
-
-const toggleUIControlPattern: RegExp = new RegExp(
-  "[!UICONTROL " + wordMatch + "]" + wordMatch + "+"
-);
-function toggleUIControl() {
-  return surroundSelection("[!UICONTROL ", "]", toggleDNLPattern);
-}
+export const wordMatch: string = "[A-Za-z\\u00C0-\\u017F]";
+export const urlRegex: RegExp = /^(http[s]?:\/\/.+|<http[s]?:\/\/.+>)$/;
