+++ src/lib/commands/toggle-dnl.ts
@@ -1,0 +1,10 @@
+import { wordMatch } from "../commands";
+import { surroundSelection } from "../editorHelpers";
+
+const toggleDNLPattern: RegExp = new RegExp(
+    "[!DNL" + wordMatch + "]" + wordMatch + "+"
+);
+
+export function toggleDNL() {
+    return surroundSelection("[!DNL ", "]", toggleDNLPattern);
+}
