+++ src/lib/commands/toggle-headers.ts
@@ -1,0 +1,26 @@
+import { surroundSelection } from "../editorHelpers";
+
+const headerWordPattern = /#{1,6} .+|.+/;
+export function toggleTitleH1() {
+    return surroundSelection("# ", "", headerWordPattern);
+}
+
+export function toggleTitleH2() {
+    return surroundSelection("## ", "", headerWordPattern);
+}
+
+export function toggleTitleH3() {
+    return surroundSelection("### ", "", headerWordPattern);
+}
+
+export function toggleTitleH4() {
+    return surroundSelection("#### ", "", headerWordPattern);
+}
+
+export function toggleTitleH5() {
+    return surroundSelection("##### ", "", headerWordPattern);
+}
+
+export function toggleTitleH6() {
+    return surroundSelection("###### ", "", headerWordPattern);
+}
