+++ src/lib/commands/toggle-success.ts
@@ -1,0 +1,16 @@
+import { newLine } from "../commands";
+import { surroundBlockSelection } from "../editorHelpers";
+
+const startingSuccess: string = `>[!SUCCESS]${newLine}>${newLine}>`;
+const endingSuccess: string = newLine;
+const successBlockWordPattern: RegExp = new RegExp(
+  startingSuccess + ".+" + endingSuccess + "|.+",
+  "gm"
+);
+export function toggleSuccess() {
+  return surroundBlockSelection(
+    startingSuccess,
+    endingSuccess,
+    successBlockWordPattern
+  );
+}
