+++ src/lib/commands/toggle-uicontrol.ts
@@ -1,0 +1,10 @@
+import { wordMatch } from "../commands";
+import { surroundSelection } from "../editorHelpers";
+
+const toggleUIControlPattern: RegExp = new RegExp(
+  "[!UICONTROL " + wordMatch + "]" + wordMatch + "+"
+);
+
+export function toggleUIControl() {
+  return surroundSelection("[!UICONTROL ", "]", toggleUIControlPattern);
+}
